dangerouslySetInnerHTMLをuseEffectに置き換えてセキュリティリスクを解消
