+++ src/components/NikeLog.tsx
@@ -86,197 +86,6 @@ export const NikeLog = ({ summaries, shuffledImageNumbers }: NikeLogProps) => {
   // 日付単位のメトリクスを計算
   const dailyMetrics = calculateDailyMetrics(summaries)
 
-  // Chart.jsの初期化用のスクリプトタグを生成
-  const initScript = `
-    window.addEventListener('DOMContentLoaded', function() {
-      // 日付表示用の関数をクライアントサイドで定義
-      const formatDate = (date) => {
-        return new Date(date).toLocaleDateString('ja-JP', {
-          month: '2-digit',
-          day: '2-digit'
-        }).replace('/', '/');
-      };
-
-      // フィルタリング関数の定義
-      const filterByMonths = (data, months) => {
-        const now = new Date();
-        return data.filter(m => {
-          const date = new Date(m.date);
-          const monthDiff = (now.getFullYear() - date.getFullYear()) * 12 + now.getMonth() - date.getMonth();
-          return monthDiff <= months;
-        });
-      };
-
-      // 初期データの設定
-      const rawData = ${JSON.stringify(dailyMetrics)};
-      
-      // 画面サイズに応じたデータのフィルタリング
-      const getFilteredData = () => {
-        if (window.innerWidth >= 1024) {
-          return filterByMonths(rawData, 4);  // PC: 4ヶ月
-        } else if (window.innerWidth >= 768) {
-          return filterByMonths(rawData, 3);  // タブレット: 3ヶ月
-        } else {
-          return filterByMonths(rawData, 1);  // スマホ: 1ヶ月
-        }
-      };
-
-      // グラフの初期化
-      const canvas = document.getElementById('dailyMetricsChart');
-      if (!canvas || typeof Chart === 'undefined') return; // Chart.js 未ロード対策
-      
-      const ctx = canvas.getContext('2d');
-      let currentData = getFilteredData();
-      
-      const chart = new Chart(ctx, {
-        type: 'line',
-        data: {
-          labels: currentData.map(m => formatDate(m.date)),
-          datasets: [
-            {
-              label: 'セッション数',
-              data: currentData.map(m => m.sessions),
-              borderColor: '#4ECDC4',
-              tension: 0.1,
-              yAxisID: 'y1',
-              borderWidth: 2,
-              pointRadius: 0,
-              pointHoverRadius: 4
-            },
-            {
-              label: 'メッセージ数',
-              data: currentData.map(m => m.messages),
-              borderColor: '#FF6B6B',
-              tension: 0.1,
-              yAxisID: 'y2',
-              borderWidth: 2,
-              pointRadius: 0,
-              pointHoverRadius: 4
-            },
-            {
-              label: 'リピートユーザー数',
-              data: currentData.map(m => m.repeats),
-              borderColor: '#FFD93D',
-              tension: 0.1,
-              yAxisID: 'y1',
-              borderWidth: 2,
-              pointRadius: 0,
-              pointHoverRadius: 4,
-              borderDash: [5, 5]
-            }
-          ]
-        },
-        options: {
-          responsive: true,
-          maintainAspectRatio: false,
-          interaction: {
-            mode: 'index',
-            intersect: false,
-          },
-          plugins: {
-            legend: {
-              position: 'top',
-              align: 'start',
-              labels: {
-                boxWidth: 15,
-                padding: 15
-              }
-            },
-            tooltip: {
-              callbacks: {
-                afterBody: function(context) {
-                  const index = context[0].dataIndex;
-                  const data = currentData;
-                  const sessions = data[index].sessions;
-                  const repeats = data[index].repeats;
-                  const repeatRate = sessions > 0 ? ((repeats / sessions) * 100).toFixed(1) : '0.0';
-                  return \`リピート率: \${repeatRate}%\`;
-                }
-              }
-            }
-          },
-          scales: {
-            x: {
-              grid: {
-                display: true,
-                color: 'rgba(255, 255, 255, 0.1)'
-              },
-              ticks: {
-                maxRotation: 45,
-                minRotation: 45,
-                padding: 5,
-                color: 'rgba(255, 255, 255, 0.8)',
-                font: {
-                  size: 11
-                }
-              }
-            },
-            y1: {
-              type: 'linear',
-              display: true,
-              position: 'left',
-              title: {
-                display: true,
-                text: 'セッション数・リピート数',
-                color: '#4ECDC4',
-                font: {
-                  size: 12
-                }
-              },
-              grid: {
-                color: 'rgba(255, 255, 255, 0.1)'
-              },
-              ticks: {
-                color: 'rgba(255, 255, 255, 0.8)',
-                padding: 8,
-                font: {
-                  size: 11
-                }
-              }
-            },
-            y2: {
-              type: 'linear',
-              display: true,
-              position: 'right',
-              title: {
-                display: true,
-                text: 'メッセージ数',
-                color: '#FF6B6B',
-                font: {
-                  size: 12
-                }
-              },
-              grid: {
-                display: false
-              },
-              ticks: {
-                color: 'rgba(255, 255, 255, 0.8)',
-                padding: 8,
-                font: {
-                  size: 11
-                }
-              }
-            }
-          }
-        }
-      });
-
-      // リサイズイベントの処理
-      let resizeTimeout;
-      window.addEventListener('resize', function() {
-        clearTimeout(resizeTimeout);
-        resizeTimeout = setTimeout(function() {
-          currentData = getFilteredData();
-          chart.data.labels = currentData.map(m => formatDate(m.date));
-          chart.data.datasets[0].data = currentData.map(m => m.sessions);
-          chart.data.datasets[1].data = currentData.map(m => m.messages);
-          chart.data.datasets[2].data = currentData.map(m => m.repeats);
-          chart.update();
-        }, 250);
-      });
-    });
-  `;
-
   return (
     <div id="category-nikelog" className="category-content block" data-content="nikelog">
       {/* 全期間の合計統計 */}
@@ -303,7 +112,7 @@ export const NikeLog = ({ summaries, shuffledImageNumbers }: NikeLogProps) => {
         <div className="max-w-[2000px] mx-auto">
           <h2 className="text-xl font-bold p-4 text-white">日別利用状況</h2>
           <div className="h-[450px] px-2 pb-4">
-            <canvas id="dailyMetricsChart"></canvas>
+            <canvas id="dailyMetricsChart" data-metrics={JSON.stringify(dailyMetrics)}></canvas>
           </div>
         </div>
       </div>
@@ -377,7 +186,6 @@ export const NikeLog = ({ summaries, shuffledImageNumbers }: NikeLogProps) => {
           </div>
         ))}
       </div>
-      <script dangerouslySetInnerHTML={{ __html: initScript }} />
     </div>
   )
 }
